TDM and DM should now be working properly GetScore returns a vector sized after the number of players

+++ Project/src/game/gamemodes/deathmatch/Deathmatch.h
@@ -14,7 +14,7 @@ public:
 	virtual int checkWin();
 
 private:
-	size_t m_maxKills;
+	float m_maxKills;
 	int m_winnerID;
 
 };
